Follow the 'connections' uri when present
* Full access to the API requires both user consent and designating
which organizations the user wants the application to be able to
interact with. Follow the 'connections' uri if present on any org to
allow the user finish the setup process.
* The connections app knows how to issue a redirct when the
select-organization route is opened with a `redirect_uri` query parameter.
This can send the user back to the original app once they have completed
the organization selection process and makes exercising the api through
the demo app smoother.